Gate test utilities for tests only
