+++ src/lib.rs
@@ -7,7 +7,7 @@
 pub mod language;
 pub mod parser;
 pub mod syntax_utils;
-#[cfg(any(test, feature = "test-util"))]
+#[cfg(test)]
 #[doc(hidden)]
 pub mod test_util;
 pub mod tokenizer;
